requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
click==8.1.8
coloredlogs==15.0.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
flatbuffers==25.2.10
greenlet==3.2.2
h11==0.16.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==25.0
passlib==1.7.4
pillow==11.2.1
protobuf==6.31.1
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
